+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and
 this project adheres to [Semantic Versioning](https://semver.org).
 
 
+## [0.3.0] - 2019-10-25
+
+Now available as rez package.
+
+### Added
+
+- `package.py` for [rez package](https://github.com/nerdvegas/rez).
+- rez package install CI for Linux and OSX.
+- Windows (bash) tests for CI
+
+
 ## [0.2.0] - 2019-10-25
 
 ### Added